l2

+++ Controllers/UsersController.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -10,6 +7,7 @@ using praktika.Models;
 
 namespace praktika.Controllers
 {
+    [Authorize(Roles = "Admin")]
     public class UsersController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -19,43 +17,29 @@ namespace praktika.Controllers
             _context = context;
         }
 
-        // GET: Users
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Users.Include(u => u.Role);
-            return View(await applicationDbContext.ToListAsync());
+            var users = await _context.Users.Include(u => u.Role).ToListAsync();
+            return View(users);
         }
 
-        // GET: Users/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _context.Users
-                .Include(u => u.Role)
-                .FirstOrDefaultAsync(m => m.IdUser == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var user = await _context.Users.Include(u => u.Role).FirstOrDefaultAsync(m => m.IdUser == id);
+            if (user == null) return NotFound();
             return View(user);
         }
 
-        // GET: Users/Create
+        [AllowAnonymous]
         public IActionResult Create()
         {
             ViewData["IdRole"] = new SelectList(_context.Roles, "IdRole", "Name");
             return View();
         }
 
-        // POST: Users/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
+        [AllowAnonymous]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("IdUser,Login,Password,Email,IdRole")] User user)
         {
@@ -63,41 +47,26 @@ namespace praktika.Controllers
             {
                 _context.Add(user);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Index", "Home");
             }
             ViewData["IdRole"] = new SelectList(_context.Roles, "IdRole", "Name", user.IdRole);
             return View(user);
         }
 
-        // GET: Users/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var user = await _context.Users.FindAsync(id);
-            if (user == null)
-            {
-                return NotFound();
-            }
+            if (user == null) return NotFound();
             ViewData["IdRole"] = new SelectList(_context.Roles, "IdRole", "Name", user.IdRole);
             return View(user);
         }
 
-        // POST: Users/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("IdUser,Login,Password,Email,IdRole")] User user)
         {
-            if (id != user.IdUser)
-            {
-                return NotFound();
-            }
-
+            if (id != user.IdUser) return NotFound();
             if (ModelState.IsValid)
             {
                 try
@@ -107,14 +76,8 @@ namespace praktika.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!UserExists(user.IdUser))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!UserExists(user.IdUser)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -122,36 +85,20 @@ namespace praktika.Controllers
             return View(user);
         }
 
-        // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _context.Users
-                .Include(u => u.Role)
-                .FirstOrDefaultAsync(m => m.IdUser == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var user = await _context.Users.Include(u => u.Role).FirstOrDefaultAsync(m => m.IdUser == id);
+            if (user == null) return NotFound();
             return View(user);
         }
 
-        // POST: Users/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var user = await _context.Users.FindAsync(id);
-            if (user != null)
-            {
-                _context.Users.Remove(user);
-            }
-
+            if (user != null) _context.Users.Remove(user);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
